requirements.txt
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
scipy>=1.9.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0
pyyaml>=6.0
biopython>=1.79
python-Levenshtein>=0.12.0
joblib>=1.1.0
tqdm>=4.64.0